Interval Search Tree Implementation
Implementation uses an Un-balanced BST.
Currently supports the following methods.
- insertion
- low_keys
- high_keys
- max_keys

The *_keys API is pretty ugly at the moment.  Might clean-up later but not much benefit since this is code that is focused on being able to work with intervals and not necessarily full API support.

The private Node class does support a comparator implementation that is used by the insertion to handle the max point for the current parent of the sub-tree.