Chapter 7 Problem 6 Implementation
Finished addressing the collinear logic to determine the maximum number of points that exist in a line.
- There existed an issues with the `Comparator` in the `Point` class that broke the transitive laws, and therefore would throw a

```
Comparison method violates its general contract!
```

This required to change to a direct equal comparison of doubles (bit by bit equality).

The approximation is however maintained with the actual collinear logic since the comparator is only used for sorting.