[FTR] Alphabet Algorithm Support
Feature Description

This feature development is based off of the comment within the pull
request to support Alphabet keying for the D words list and iterate
through the characters of S managing the Node of word W and current
index for said word.  Opposite of mapping in which just S was mapped.
This time S is iterated once, and all D words stored in a tuple with
the word and current index. When an index character matches the tuple
index is updated by one and then added back to its new character tuple.

Related Issues
- #2
- #3