[FTR] Challenge 2 - Word Count Engine
Feature Description

Want to record each occurrence of a word contained within a document.
The occurrence count then should be incharge of the order in which the
result list is returned along with the count of the word itself.  The
Expected word is to operate only on ASCII characters and a word is
described as being only lower or upper case.

Parsing of puncutation is needed to be performed during parsing.
If two words have the same hit count in the document then the order
returned should be the word that occurred first is shown first.

Related Issues
- #7