[FTR] Challenge-3 Minimum Sales Path
Feature Description

Minimum Sales Path problem with a k-d tree is implemented with depth
traversal across N-branches of a parent node.  Similar to what most
serial tree traversals look like.

This implementation uses data agumnetation due to the fact dont know
if the referenced data is clean for modification or not.

Related Issues
- #11