[FTR] Challenge-4 Spell Checker Trie
Feature Description

Implementation of a Trie to support a spell checker that is capable of
ignoring cpaitizalization and mismatch vowels.  The Trie is definietly
more error prone, larger performance negative, and memory space complexity.
However it might provide a way to get more words that are 'close' to the
mismatch.

Related Issues
- #13