[FTR] Challenge-10 Recursive Count Analysis
Challenge-10 [A]nalysis portion of the the recursive implementation.
The analysis shows that it has suboptimal solution and has subproblems
that reoccurr.

Therefore this can be addressed by dynamic programming.  In addition,
the runtime performance is exponential while the memory impact is linear
or equal to the height of the tree.

Related Issues
- #23