[FTR] Challenge-9 Subproblem Identification & Memoization
Feature Description

Breaking the problem to still use recursion but now be responsible for
memoization of the results.  If the data didnt already exist it will
progress down the branch necessary to populate the data that is missing.

Once it has been populated, the populated value will be used instead of
traversing down the branch.

Related Issues
- #22