[FTR] Challenge-11 Turn Upside Down
Feature Description

Changed the recursive problem into a serial problem for the dp problem.  At
this point in time now the run time cost is O(M * N) where N is the number
as the number of elements equal to the value and M is the size of the coin
set.  More could be done to aid in reducing this but is beyond scope that just
addressing the dp implementation.

Related Issues
- #26