[FTR] Challenge-12 [A]nalysis of Max SubSquareSize Matrix
Feature Description

The analysis for the max subsquare size contained within a given matrix
looks expoential with the power factor being 3 and the base being the
maximum of the length vs width of the matrix itself.  This relates to
the maximum height of the tree that is possible.

Related Issues
- #27