[FTR] Challenge-12 [S]ubproblem Inspection - Max subsquare in Matrix
Feature Description

The following performs the memoization of the max and values seen for
a given position within the matrix.  It still is recursive so it is top
down and builds back up.  The check if a value has already been computed
for a given row / col is based on if it is set to zero for the memoization
at the given row / col.  If the cell is false then the value will be set
to `Integer.MIN_VALUE`.

Related Issues
- #27