[FTR] Challenge-12 [T]urn the Problem Around
Feature Description

Challenge-12 Max subsquare contained in the matrix implementation is now
serially done instead of recursively.  The impact is that the time is now
only O(N*M) where N is the length and M is the width of the input matrix.

Memory impact is O(2*(M*N)) due to having a max and value memoization values
being stored.  This can be reduced to jsut the length of the row, since the
problem only processes one row at a time.

Related Issues
- #27