[FTR] Challenge-13 [T]urn the Problem Around - Knapsack 0/1
Feature Description

[T]urn the problem around from a recursive to a serial problem.  There
is actually a bit different in going from recursive to serial not just
due to the wording but how you have to look at specific elementing within
the cache itself.

For a current value with a specified weight will include it if it is the
max if its value plus the value at the previous col's weight that remains
is larger than the previous value for the given column.

Related Issues
- #28