[FTR] Chapter 9 Problem 1 - [F]unctional Implementation
Feature Description

Completeion of the functional implementation for determining the total
number of ways in which the steps can be traversed.

Key to the logic is that for the top down is that step N does not have
to be the last step prior to the end and thus need to also iterate on
step N-1, and N-2 steps.

This means that the overal recursive method actually provides on the
count of unique traversals from the provided top step from step any
allowing for single, double or triple steps.

Related Issues