[FTR] Chapter 9 Problem 2 - [F]unctional Implementation
Feature Description

Recursive implementation that handles calculating the total number
of unique paths there are from the origin to the designated cell.

In this case the origin is 0,0 while the destination is the row,col
of the provided matrix.

The current implementation does not consider the occurrence of blocked
locations.

Related Issues