[FTR] Chapter 9 Problem 3 Recursive Implementation
Feature Description

Implementation of the recursive definition.  This is in contrast to the
serial implementation which reduces the memory impact but the code is
significantly more error prone.

The memory impact due to the stack will be O(lgN) where N is the length
of the input array.

Related Issues