[FTR] Chapter 9 Problem 4 - [A]nalysis
Feature Description

Breakout of stating that the problem in its current form does not have
optimal substructure due to the dynamic creation of a list in each prior
step in order to maintain current state.  The overall time therefore will
be worse than say just computing the total numnber of possible subsets.

This does not mean the problem can not be implemented iteratively, however
the memory impact will still be dominated by the O(2^N *N) for the set
expanisions.

Related Issues