[FTR] Chapter 9 Problem 5 - [A]nalysis
Feature Description

Analysis step of the problem.  Due to stateful elements existing within
each recursive layer, the problem / recursive implementation is likely to
not benefit from the traditional optiizations of dynamic programming in
its current form.

It is possible to implement a serial implementation to save on the
recursive method overhead, however the permutation set will likely grow
greater since it is not linear unlike the height of the problem.

Related Issues