[FTR] Chapter 9 Problem 8 - [S]ubproblem Identification & Momeization
Feature Description

Part one of caching the values for determining the number of different
ways to provide change back for a given value.  At the moment there
is a bug in the caching amount. Used dollar as base and increased the
supported denominations to be returned.  Expected is 293 and getting
38 back.  So point in time of caching is likely incorrect.

Related Issues