[FTR] Chapter 9 Problem 9 - [F]uncational Implementation
Feature Description

Recursive implementation for the N-Queen problem that will print out the
board with the position of each queen iff it is a valid location.

The implementation was performed using a HashMap to HashSet that provides
the available positions for the queen.  Allows to not have to perform full
board inspection for removing or checking if a position is available.  Just
iterate on the next position for the given row.

Related Issues