[FTR] Chapter 11 Problem 2 - Implementation
Feature Description

Implementation of the challenge problem.  Implemented using a combination
of HashMap & TreeMaps along the principle of bin sorting based on a computed
key value (in this case the anagram).  The generate anagram is a consistent
key for two strings that are an anagram of each other.

Related Issues
- #51