[FTR] Chapter 11 Problem 3 - Implementation
Feature Description

Implemented a modified binary search that must examine the current index
and low and high index values in the event that a match is not found.

Primary just index logic to go the correct direction even in the event
of random rotation of the sort array.

Related Issues
- #52