[FTR] Chapter 11 Problem 3 - README Issue & Solution Notes
Feature Description

Notes on issues and solution aspects of the problem.  The problem itself
is quite tricky for handling the duplicate case in which duplicate spans
both the low mid and high indices since can not indicate where the break
lies.

This requires more of a recursive implementation or a BFS saving of data.

Related Issues
- #52