[FTR] Chapter 11 Problem 5 - Implementation
Feature Description

Provided a recursive implementation that will perform two branches when an
empty string is encountered when performing the search.  In the overall worse
case this can lead to a O(N) run time performance along with a O(lgN) memory impact
due to the stack usage.

Related Issues
- #54