[FTR] Challenge 16 - BST Successor Node
Feature Description

Implementation to support obtaining the successor node within a given
binary search tree at a specific node.  Since this is not a balanced BST
this means that the worse case execution time that would be expected is
O(N).  Memory impact of just the search is O(C) due to the fact that no
additional memory footprint is expended.

Related Issues
- #65