[FTR] Chapter 11 Problem 8 - Implementation
Feature Description

Implementation of the non-balanced binary tree for the time being.  This
means that the worse case for finding the rank for a given number will be
O(N).

Duplicates will always occur on the left side and will factor into the count
for a rank calculation.

Related Issues
- #57