[FTR] Challenge 18 - Max Binary Tree
Feature Description

Implementation of the max binary tree that is constructed out of an array
of integers.  This implementation has a worse case expected run time of
O(N^2) due to the worse case arrangement of a sorted array.

A sorted array will require a full N + (N - 1) + (N - 2) + (N - 3) + ... + 1
number of searches for the max.  This is therefore an O(N^2) time.  In terms
of memory performance expect that the maximum stack depth is also in the worse
case in which the tree is not balanced resulting in a memory impact that is
linear on the stack O(N).

Related Issues
- #75