[FTR] Challenge 18 - Stack Based Solution for Max Binary Tree
Feature Description

The serial implementation only iterates through the array of nums once by
using a stack.  The stack will contain the largest number at the bottom
in the form of the node, and each root node value on top must be smaller
and come after the one below it in the array.

If the new node that is created is larger than the top of the stack pop
it off into a temporary tree.  Keep poping off as long as the new node value
is greater than the current top of the stack.  As nodes are popped off
they will act as the parent to the previous node that was previous popped off
with the child being on the right.  Once a value that is equal or greater
or the stack is empty, the place the last popped off node as the new node's
left and push the new node onto the stack.

At the end empty the stack with each popped off being the child of the next
layer but to the right.

Another important aspect is that the order of the nodes in an inorder
traversal shall result in traversal of values that is equal to the order
of the values in the array.

Related Issues
- #75