[FTR] Lesson 2 - Problem 4 BST Insertion Orders
Feature Description

Development of the population of the BST insertion orders of the
elements.  This is a fairly bit of code due to the inter-mixing of both
the left and right branches that can happen.  Expecting worse case memory
and time to be something like O(N^2*2^N) but did not look into detail for
the copying and permutation building of the arrays.

Related Issues
- #77