[FTR] Challenge 22 - Longest Word for Given Prefix Linear Implementation
Feature Description

Given a list of words and then a prefix, find the longest word with the supplied
prefix.  This has been implemented with a worse case run time of O(N) w.r.t the
length of the prefix.

Given a prefix, the trie is traversed down until all the characters of the prefix
have been examined.  If the none of the `Node` along the path are null then the
last node will be used to return the max lenght seen string.

Related Issues
- #86