[BUG] Lesson 2 Problem 3 - Towers of Hanoi Solution
BUG Description

Modification to the previous attempt to correct based on what is actually
desired.  Copied the solution from the course.  Need to brake down that
all disks above the target need to be moved from source to aux.  Then
final disk to destination, then all others from aux to destination in the
simplest of terms.

Related Issues
- #77