[FTR] Lesson 3 Problem 4 - Edit Distance Implementation
Feature Description

Recording all the different set of moves that can exist in the minimum
number of operations on s1 with respect to s2.  The tree is trimmed in
such a way that once the number of operations is greater than the
minimum found it will not longer process the other operations.

This allows for supporting arbitrary insertions and deletes at any
index which may be desired due to insertion could allow for all other
letters aligning correctly.

Related Issues
- #89
- #97