[FTR] Challenge 24 - Array Volume Implementation
Feature Description

Linear implementation that details keeping the maximum index and
storing a decreasing set of values on a stack along with their
indices.  When a value is found to be greater that the max then
will process the entire stack.  If find a value that is less need
to pop off all the values less than it and update the volume along
the way.

Should only ever see two indices twice due to putting on the stack
potentially and then popping off for volume calculation.  Worse case
memory impact expected would be lienar with the input (a decreasing
set of heights).

Related Issues
- #104