[FTR] Lesson 4 Problem 3 - Knapsack Weight/Value/Quantity Implementation
Feature Description

Implementation of the knapsack problem that is balancing the mix of
weight, value, and quantity with respect to the items available.  Will
always take the max value that is available at the lowest weight possible.

If multiple values are the same and at the same weight, under the maximum
that can be carried then both lists are accepted.

Related Issues
- #106
- #109