[FTR] Inheritance
Feature Description

Basic inheritance between two different classes.  Takes into account
the separated header files with the classes and defined namespace
definitions for each.

At the moment there is a warning since reference a direct string for
a `char *` but this is okay since just doing a full copy for the name.
If attempted to perform a modification on this address will throw a
run time exception since this is not modifiable.

Related Issues
- #130