[FTR] Class Copy Constructors
Feature Description

Prototyping of class constructor for the copy type.  Child class relies
on default parent constructor by default.  Copy implementation is good
for when the class has more complex attributes in which need to be tied
to a specific instance.  Can determine how much of the class to copy over
full vs shallow.

Related Issues
- #130