[FTR] Virtual Class Members
Feature Description

Deconstructor of base class along with change to abstract out the child
classes of the Person  class.  This is more oriented towards a factory
design such that the user call of the Person library can wrap the Person
class in the construction to aid in direct dependencies.

Related Issues
- #130