[FTR] Chapter 2 Problem 2 - Kth Element from Last
Feature Description

Implementation of the Kth element from last in a fictious linked list
class construct.  This time used templated types for the linked list
class to work with.

This is a classic two pointer linked list problem in which you move
ahead by the kth amount - 1 first.  If you run out or dont get the
count to be one then the linked list is not big enough.

Related Issues
- #152