[PR] Chapter 2 Problem 4 - Partition
Feature Description

Implementation of the partition method in O(N).  The partition requires
inspection of all elements of the linked list.  Once again using a fake
linked list class to operate directly on the next and data attributes.

This time since performing a comparison the generic must extend from the
`Comparable` interface to gain access to the `compareTo` function.

Managing multiple heads and multiple tails.  Break down the next pointer
for each node then reset it once it is determined which bucket it falls
within.  For the second partition half this means that the break was
likely un-necessary but is done to keep the logic consistent.

Related Issues
- #152