[FTR] STL List Prototyping
Feature Description

Working with the basic list type which as defined in the docs is a
doubly linked list implementation.  Can use some of the same templated
functions as vector but not operations are supported (i.e. indexing [],
and `.at(index)`).

For specific on insertion need to use iterator that indicates where the
insertion is going to happen.  Insert does not perform the traversal only
the modification of the node pointers.

Related Issues
- #161