[FTR] Chapter 3 Problem 2 - Stack Min
Related Issues

Two forms of implementation, one that has class augmentation to hold
the minimum for each new value placed onto the stack.  The second only
records the minimum in a separate stack such that each new value does
not need to have the space for it.  If a new min that is equal to or
less than the previous is added then add the new min to the min stack
upon a push on the overall data stack.

Related Issues
- #163