[FTR] Chapter 3 Problem 3 - Stack of Plates
Feature Description

Implementation to handle over-flow of stack elements once a capacity
amount is reached.  Also supports popping values from a specific
index, as long as the index is supported.

Not going to support balanced pushing with the `popAt` function since
is does subtract from the original push that would be equal to the
original stack functionality.

Related Issues
- #163