[FTR] Chapter 3 Problem 6 - Animal Shelter
Feature Description

Implementation that provides a FIFO on the animal shelter set of
cats and dogs only such that a person can add either.  But for removing
must get the first, or the first dog or cat.

This uses two queues in the form of a linked list, one for cats and
one for dogs.  The way to get the next available is based on the
index associated with it.

If wanted to do it all within a single linked list implemenation
would need to be able to manage the out or order next and reconnecting
the previous.  Using a counter will just need to deal with the roll
over but dont know if a shelter would need to deal with 2.14 B animals.

Related Issues
- #163