[FTR] Chapter 4 Problem 2 - Minimal Height Tree
Feature Description

Creation of a minimal binary search tree from a given sorted array.
Using a recursive method to easily break up the left and right node
creation for a given middle index which is acting as the root.

O(N) for execution time since all indices of the given data array is
visited once.  In terms of memory impact besides the creation of the
data structure itself, will see an additional O(lgN) for the stack
space due to the maximum height of a balanced tree is always O(lgN).

Related Issues
- #170