[FTR] Chapter 4 Problem 7 - Build Order
Feature Description

DFS implementation of the build order problem.  There is also a BFS solution
as well but requires building a graph and knowing the root nodes, and then
removing nodes from the graph as they are not needed.

This is all done by managing the number of outgoing / incoming connections
for a given graph node.

The recusive is quite simple in terms of just trimming the current hashset
used to manage the DFS route.  When going down a route if the current dependency
already exists within teh hashset then there will exist a circular loop.

Will add to the hashset when going down.  Will pull in all the dependencies for
a given one at one specific point.

Related Issues
- #170