[FTR] Chapter 4 Problem 9 - BST Sequences
Feature Description

Permutation / combinatrics problem w.r.t a BST.  Pretty complex but a good
problem to practice on.  For this problem the basic comes down to building
the list of lists of integers from the bottom up.  This is due to the fact
that a parent must always come before the children  that are under it.

Permutation / combinatrics is needed due to address possible insertion order
across parents at the same level and their respective children can occur at
any point in time.

THe top level ends up being a DFS traversal, which can be done recursively
or iteratively with the children for each node populating the list of lists
for the parent.

Then just perform the listPermutations function that does the list merging
along with the parent value.

Related Issues
- #170