[FTR] Chapter 4 Problem 11 - Random Node
Feature Description

Obtaining a random node in constant time is achieved by using the random
operation on the total number of nodes within the binary tree.  The nodes
are given an index to within a list that has the given node.

In order to support the insert and delete operations as well with this,
the insertion just adds the new node to the back of the arraylist.  Additionally
a hashmap of the node to the index is also updated.  On delete this means
that will performa a swap of the node at its given index with the last node
in the array list.  Then update the index within the map for the swapped
entry.

Then the last element of the array is removed along with the node from the
hashmap.

Related Issues
- #170