[FTR] Chapter 4 Problem 12 - Paths with Sum
Feature Description

Uses dynamic programming to record the previously seen running sums such
that is they are minused with target then have a valid path so increment.

Keep a map of the running sum along with the number of occurrences total
in the current DFS traversal.  This means prior to branching add it with the
total number of occurrences + 1.  Prior to leaving for the current node
subtract the number of occurrences for the running sum by 1.  If zero remove
the key from the map entirely.

Related Issues
- #170