[FTR] Challenge 37 - Island Count Implementation
Feature Description

Serial implementation of the counting of the number of islands in the
given map.  The does require extra memory (implementation with copy of
number of rows).

The implementation does restrict performance to that of only being equal
to a O(N*M) for the traversal of the map itself.

Related Issues
- #179