[FTR] Challenge 37 - Recursive Implementation
Feature Description

Providing of recursive implementation to provide an testing base for correctness
since it is easier to implement correctly.

Additionally provided some generation of random data sets to test both implementations
and will now error if there is a difference between the two.

Current there is a bug to to the reduction of unique value not occurring just yet.

Related Issues
- #179