[BUG] Challenge 37 - Island Count Set
BUG Description

So just holding onto unique value set is not good enough.  What really needs to
happen is have a graph of the values that point to a specific root to resolve the
total number of unique roots within the graph is all that I care about.

Related Issues
- #179