[FTR] Challenge 37 - Island Count Serial Modification
Feature Description

Just recording the available key is not enough.  There are cases in which
you need to know which keys are still available but the two being compared
may or may not be part of that set.  It might just be that as the island
layout progresses that the available unique IDs are not the ones being worked
with.

Therefore need a set association with the set of unique island ids that are
currently being worked upon to acheive single row/col access.  Implemeneted a
set of mapping unique generated island ID to its root, initially set to itself.

When a left and top node differ, then will pull the root for both.. Then randomly
select which root node will be put into the dp memory, and update the one that
was not to now point to the one that was assigned.

This forces that the worse case requires a lookup of the parent of the unique id.
The cost is proportional to the height of the graph.  The implementation is trival
such that only selection is done randomly.

Related Issues
- #179