[FTR] Challenge 38 - Water Volume Stack Implementation
Feature Description

Stack based implementation that is responsible for calculating the total volume
based on the array of heights.  The key here is handling the empty list after
the pop of prev height, and pushing on the correct index after the fact which
indicates the current water level that has been accounted for.

Related Issues
- #180