[FTR] Challenge 38 - Water Volume Two Scan Implementation
Feature Description

Handling of the two scan method that records the maximum height that
is seen based on direction.  THis is the greedy approach.

Once the maximums based on both directions are computed then can determine the
volume at each index based by computing the min of the maxs then subtracting from
the current height.

Related Issues
- #180