[FTR] Chapter 8 Problem 1 - [Er]adicate Memory Triple Step
Feature Description

From the development of the serial dp implementation along with the data
from the tests it can be seen that a memory opitimization can occur to
only record the last three steps. This is done in a queue against a
total value.  The total is added for each new step, then total is incremented
by the value amount added minus the value at the front of the queue.

Related Issues
- #183