[FTR] Chapter 8 Problem 3 - [F]unctional Magic Index
Feature Description

Recursive definition for finding the magic index.  The key here is that the
data is sorted, and the that values can be anything as long as they are sorted.

When finding a value that is negative then dont need to keep looking left.  if
find a value that is greater or equal to the length of the array dont need to
look right anymore either.

If a number is within the range then can see if the value at the index is
less than current index.  If so can perform direct jump to that index.
If number is greater than current index then perform direct jump to that
index.  However range checks of min and max index for the current level
must be respected to prevent infinite loops in addition to touching a
specific index more than once.

Related Issues
- #183