[FTR] Chapter 8 Problem 3 - [A]nalysis Magic Index
Feature Description

Providing the analysis for the magic index problem along with a brute force
solution to check the result of the problem.

The recursive solution only finds a single answer not all of them, therefore
the brute force implementation for testing records all the results and checks
that the recursive result exists within the set.

Related Issues
- #183