[FTR] Chapter 4 Problem 12 - Path Sum
Feature Description

The previous solution only considers from root, to any node.  It does not
consider the start of the downward sum starting from any node in the tree.

In order to gain this, it is required to start a new recursive call with
the current node as the new root of the tree with an empty map.  A new
recursive call should be started if the current map is not empty (will be
empty when the root node of the tree removes its value from the hashmap).

Related Issues
- NA