[FTR] Chapter 8 Problem 5 - Recursive Multiply with DP Caching
Feature Description

Recursive implementation with dp caching.  This implementation does
not modify the value in operand a, only the value in operand B.  What
happens is that the check for even / odd B value occurs like the previous
implementation but this time now will only divide the value of B by two
as a parameter for the recursive call.

The computed result is stored in the cache.  Now does break down into
pairs due to the way it is computed but is still a bit un-neccessary.

Related Issues
- #183