[FTR] Chapter 8 Problem 6 - Towers of Hanoi
Feature Description

Recursive implementation for moving data between different towers (which are
represented as stacks).  Displaying the destination stack at the end along
with the total number of steps that occurred.  Total number of steps is
related to the number of elements which is (2^N) - 1.

At the moment having the stacks is to represent the actual movement of objects
that would be required.

Related Issues
- #183