[FTR] Chapter 8 Problem 6 - [S]ubproblem Identification Towers of Hanoi
Feature Description

Implemented a caching aspect for the towers of hanoi that focuses more on the
argument values with only keeping track of the passed in arguments.  There
might be some invalid source, aux stack values due to no always modifying the
memory location but the passed in value.

Expect that the source and destination values are correct however, along with
total number of recursive calls.  Could enforce zero of source and aux by
using a object reference that contains the count instead but is un-necessary.

Related Issues
- #183