[FTR] Chapter 8 Problem 8 - [F]uncation Permutations with Dups
Feature Description

Iterative Recursion at each level that is also responsible for managing
the characters that have been processed for each index.  Keeping this
information at each level prevents processing the same permutation when
duplicates exist.

Used a hashset of the final result to check that the results are unique.

Related Issues
- #183