[FTR] Chapter 8 Problem 11 - [T]urn Around Coins
Feature Description

Serial dynammic programming to make the analysis a bit easier.  The
iterative recursion adds a bit of difficulty in the case logic as to
which values from the previous row should be used for calculation.

Needed to account for current value - denomination value back inspection
of the stack.

Related Issues
- #183